done ex00

+++ cpp06/ex00/istring.cpp
@@ -2,9 +2,9 @@
 #include <sstream>  // For std::istringstream
 #include <string>
 
-int stringToInt(const std::string& input) {
+float stringToInt(const std::string& input) {
     std::istringstream iss(input);  // Create an input string stream
-    int number;
+    float number;
     iss >> number;  // Extract the integer from the stream
     if (iss.fail()) {
         throw std::invalid_argument("Invalid input: not a valid integer");
@@ -18,7 +18,7 @@ int main(int ac, char **av) {
         // need to check for caracter alpha :  just we need string.back()=='f'
 
         // requirement :  + or - for begine ,  just digit , and the lase one have f or digits!!
-        int number = stringToInt(input);
+        float number = stringToInt(input);
         // same thing for float and double !!
         std::cout << "Converted number: " << number << std::endl;
     } catch (const std::invalid_argument& e) {
